charaobj: match onDrawDebug (signed CFlatCenterState compare)

+++ src/charaobj.cpp
@@ -3230,7 +3230,7 @@ int CGCharaObj::calcCastTime(int itemId)
  */
 void CGCharaObj::onDrawDebug(CFont* font, float posX, float& posY, float posZ)
 {
-	if ((m_weaponNodeFlagBits.m_prg && (CFlatCenterState() == 0)) &&
+	if ((m_weaponNodeFlagBits.m_prg && (static_cast<int>(CFlatCenterState()) == 0)) &&
 	    ((DbgMenuPcs.GetDbgFlagsRaw() & 0x80) != 0)) {
 		char text[0x100];
 		unsigned char* script = reinterpret_cast<unsigned char*>(m_scriptHandle);
